+++ src/main/java/lotto/validator/InputValidator.java
@@ -1,4 +1,4 @@
-package lotto.view;
+package lotto.validator;
 
 import java.util.regex.Pattern;
