+++ src/test/java/lotto/view/InputValidatorTest.java
@@ -2,6 +2,7 @@ package lotto.view;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import lotto.validator.InputValidator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullSource;
